requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.6
redis==4.6.0
drf-spectacular==0.26.4
pytest==7.4.0
pytest-django==4.5.2
gunicorn==21.2.0
pycryptodome==3.20.0
django-encrypted-model-fields==0.6.5
djangorestframework-simplejwt==5.3.0
django-filter==24.1
Pillow==10.2.0 